requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
distlib==0.3.1
Django==3.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-storages==1.10
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
filelock==3.0.12
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mod-wsgi==4.7.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
python3-openid==3.2.0
pytz==2020.1
razorpay==1.2.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.1.0
whitenoise==5.2.0
